create Balance objet and fix bugs

+++ src/main/java/gildedrose/shop/output/ShopOutputBoundary.java
@@ -8,5 +8,5 @@ public interface ShopOutputBoundary
 
     void displayInventory(List<ShopItemResponse> shopItemRespons);
 
-    void displayBalance(int balance);
+    void displayBalance(double balance);
 }
